Use a modified Base32 to generate the receipt number
